move package of ResourceService to resources
the service will handle all resource, not only bar
