+++ bonita-integration-tests/bonita-query-tests/src/test/java/org/bonitasoft/engine/resources/ResourceServiceQueriesTest.java
@@ -11,7 +11,7 @@
  * program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth
  * Floor, Boston, MA 02110-1301, USA.
  **/
-package org.bonitasoft.engine.bar;
+package org.bonitasoft.engine.resources;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
