+++ services/bonita-resources/src/main/java/org/bonitasoft/engine/resources/ResourcesServiceImpl.java
@@ -12,7 +12,7 @@
  * Floor, Boston, MA 02110-1301, USA.
  **/
 
-package org.bonitasoft.engine.bar;
+package org.bonitasoft.engine.resources;
 
 import java.util.HashMap;
 import java.util.List;
