+++ services/bonita-resources/src/main/resources/org/bonitasoft/engine/resources/hibernate/resources.hbm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <!DOCTYPE hibernate-mapping PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN" "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-<hibernate-mapping auto-import="false" default-access="field" package="org.bonitasoft.engine.bar">
+<hibernate-mapping auto-import="false" default-access="field" package="org.bonitasoft.engine.resources">
 
     <filter-def name="tenantFilter" condition=":tenantId = tenantId">
         <filter-param name="tenantId" type="long" />
@@ -15,7 +15,7 @@
         <property name="type" column="type">
             <type name="org.bonitasoft.engine.persistence.GenericEnumUserType">
                 <param name="enumClass">
-                    org.bonitasoft.engine.bar.BARResourceType
+                    org.bonitasoft.engine.resources.BARResourceType
                 </param>
             </type>
         </property>
@@ -31,7 +31,7 @@
         <property name="type" column="type">
             <type name="org.bonitasoft.engine.persistence.GenericEnumUserType">
                 <param name="enumClass">
-                    org.bonitasoft.engine.bar.BARResourceType
+                    org.bonitasoft.engine.resources.BARResourceType
                 </param>
             </type>
         </property>
